+++ src/components/header/index.css.ts
@@ -4,6 +4,9 @@ export const headerStyle = style({
   display: 'flex',
   alignItems: 'center',
   gap: '1rem',
+  position: 'sticky',
+  top: 0,
+  backgroundColor: 'white',
 });
 
 export const leftElement = style({});
