+++ src/pages/new/context.ts
@@ -1,0 +1,14 @@
+import { createContext } from 'react';
+import type { Dispatch } from 'react';
+
+type FormContext = {
+  totalFormData: Map<string, any>;
+  setTotalFormData: Dispatch<React.SetStateAction<Map<string, unknown>>>;
+};
+
+const FormContext = createContext<FormContext>({
+  totalFormData: new Map(),
+  setTotalFormData: () => {},
+});
+
+export default FormContext;
