--- src/pages/new/index.css.ts
@@ -1,9 +1,0 @@
-import { style } from '@vanilla-extract/css';
-
-export const header = style({
-  fontSize: '16px',
-  fontWeight: '500',
-});
-export const cardNumber = style({
-  display: 'flex',
-});
