+++ src/pages/new/steps/index.tsx
@@ -1,0 +1,4 @@
+import Step1 from './Step1';
+import Step2 from './Step2';
+
+export { Step1, Step2 };
